Feature/3 modlog

+++ src/main/java/com/sublinks/sublinksapi/api/lemmy/v3/user/models/BanPerson.java
@@ -8,7 +8,7 @@ public record BanPerson(
     String reason,
     Boolean ban,
     Boolean remove_data,
-    Integer expires
+    Long expires
 
 ) {
